ncurses: add libtinfo.so
This change adds libtinfo.so as a symlink to libncurses.so.
And some other relevant changes around it.
This is used by some programs, which will rather look at libtinfo, than
libncurses.
libncurses provides the same library functions as libtinfo as it is
currently being built, so this way both programs looking for libncurses
will work the same, and programs looking for libtinfo will also work.