python-tweepy: add python_version=2/3; fix LICENSE
The download of LICENSE as a distfile does not work w/o create_wrksrc=yes
which OTOH makes it impossible (for me that is) to build this package.
Instead fetch the LICENSE text in th post_install() function and verify
its digest which seems to change at least once per year.