mruby: build for host with clang
Trying to build for i686 native with gcc, or trying to cross build
with i686 as the host architecture, the build fails. This seems to
be a problem with either i686 gcc or the binutils (linker).

For now build with clang for native and for the host when cross
compiling to work around the problems.

TODO: find the real reason why the compiler or linker packages
for i686 are broken and switch back to using gcc.